+++ src/vs/workbench/contrib/positronNotebook/browser/PositronNotebookEditor.tsx
@@ -21,6 +21,7 @@ import { IThemeService } from '../../../../platform/theme/common/themeService.js
 import {
 	EditorPaneSelectionChangeReason,
 	IEditorOpenContext,
+	IEditorPane,
 	IEditorPaneSelectionChangeEvent
 } from '../../../common/editor.js';
 import {
@@ -43,6 +44,7 @@ import { URI } from '../../../../base/common/uri.js';
 import { isEqual } from '../../../../base/common/resources.js';
 import { EditorInput } from '../../../common/editor/editorInput.js';
 import { IInstantiationService } from '../../../../platform/instantiation/common/instantiation.js';
+import { IPositronNotebookInstance } from './IPositronNotebookInstance.js';
 
 
 /**
@@ -123,10 +125,11 @@ export class PositronNotebookEditor extends AbstractEditorWithViewState<INoteboo
 	 * https://github.com/microsoft/vscode/issues/40114).
 	 */
 	protected override computeEditorViewState(resource: URI): INotebookEditorViewState | undefined {
-		if (this.notebookInstance &&
-			this.notebookInstance.textModel &&
-			isEqual(this.notebookInstance.textModel.uri, resource)) {
-			return this.notebookInstance.getEditorViewState();
+		if (this.notebookInstance) {
+			const textModel = this.notebookInstance.textModel.get();
+			if (textModel && isEqual(textModel.uri, resource)) {
+				return this.notebookInstance.getEditorViewState();
+			}
 		}
 		return undefined;
 	}
@@ -384,4 +387,13 @@ export class PositronNotebookEditor extends AbstractEditorWithViewState<INoteboo
 		// Call the base class's dispose method.
 		super.dispose();
 	}
+}
+
+export function getNotebookInstanceFromEditorPane(editorPane?: IEditorPane): IPositronNotebookInstance | undefined {
+	if (editorPane &&
+		editorPane.getId() === POSITRON_NOTEBOOK_EDITOR_ID &&
+		editorPane instanceof PositronNotebookEditor) {
+		return editorPane.notebookInstance;
+	}
+	return undefined;
 }
